Change some ambiguous grammar.
Previously, ws separated args in a lambda value. Now we use commas,
to stop a parsing ambiguity with juxtaposition. This only came up
with type variables, but I am not willing to sacrifice them, nor can
I think of a method to disambiguate the two cases without an extra token.